Clean up fallow findings: dead code false positive, test duplication
fallow audit --base main flagged two real things on this branch:

- OrgDeleteWorkflow.run and DomainActivateWorkflow.run read as unused
  class members: they are only ever called by the Cloudflare Workflows
  runtime via WorkflowEntrypoint, never from application code, which
  fallow's static analysis has no way to see. Configured as a known
  framework entry point via usedClassMembers instead of leaving it as
  permanent audit noise.
- domains.worker.ts, orgs.worker.ts, and storage.worker.ts each had
  their own copy of overrideEnv, the TestEnv/applyD1Migrations
  boilerplate, and (for the first two) an identical adminCookie sign-in
  helper. Extracted into tests/worker/support.ts. Also extracted
  findDomainById in routes/domains.ts, shared by ensureHostname and
  probeDomain, and a small postLink test helper for the org-teardown
  write-guard tests.

Duplication in domains.worker.ts's per-test Cloudflare API mocks is
left alone: each mock returns different results for a different
scenario, so extracting it would trade test readability for a small
line-count win, and it predates this session's changes.
