+++ .fallowrc.json
@@ -2,6 +2,8 @@
   "$schema": "https://raw.githubusercontent.com/fallow-rs/fallow/main/schema.json",
   "ignorePatterns": ["scripts/seed-local.ts"],
   "ignoreDependencies": ["tailwindcss"],
+  // "run" is called by the Cloudflare Workflows runtime, not application code.
+  "usedClassMembers": [{ "extends": "WorkflowEntrypoint", "members": ["run"] }],
   "audit": {
     "gate": "new-only",
     "healthBaseline": "fallow-baselines/health.json"
